renamed 'doubleinject' package to 'command'
