+++ src/main/java/com/hubspot/dropwizard/guice/AutoConfig.java
@@ -3,9 +3,9 @@ package com.hubspot.dropwizard.guice;
 import com.google.inject.ConfigurationException;
 import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
-import com.hubspot.dropwizard.guice.doubleinject.InjectedCommand;
-import com.hubspot.dropwizard.guice.doubleinject.InjectedConfiguredCommand;
-import com.hubspot.dropwizard.guice.doubleinject.InjectedEnvironmentCommand;
+import com.hubspot.dropwizard.guice.command.InjectedCommand;
+import com.hubspot.dropwizard.guice.command.InjectedConfiguredCommand;
+import com.hubspot.dropwizard.guice.command.InjectedEnvironmentCommand;
 import io.dropwizard.Bundle;
 import io.dropwizard.ConfiguredBundle;
 import io.dropwizard.cli.Command;
