+++ src/test/java/com/hubspot/dropwizard/guice/sample/command/TestCommand.java
@@ -1,7 +1,7 @@
 package com.hubspot.dropwizard.guice.sample.command;
 
-import com.hubspot.dropwizard.guice.doubleinject.InjectedCommand;
-import com.hubspot.dropwizard.guice.doubleinject.Run;
+import com.hubspot.dropwizard.guice.command.InjectedCommand;
+import com.hubspot.dropwizard.guice.command.Run;
 import com.hubspot.dropwizard.guice.sample.HelloWorldConfiguration;
 import io.dropwizard.setup.Bootstrap;
 import net.sourceforge.argparse4j.inf.Namespace;
