+++ src/test/java/com/hubspot/dropwizard/guice/sample/command/TestConfiguredCommand.java
@@ -1,7 +1,7 @@
 package com.hubspot.dropwizard.guice.sample.command;
 
-import com.hubspot.dropwizard.guice.doubleinject.InjectedConfiguredCommand;
-import com.hubspot.dropwizard.guice.doubleinject.Run;
+import com.hubspot.dropwizard.guice.command.InjectedConfiguredCommand;
+import com.hubspot.dropwizard.guice.command.Run;
 import com.hubspot.dropwizard.guice.sample.HelloWorldConfiguration;
 import io.dropwizard.setup.Bootstrap;
 import net.sourceforge.argparse4j.inf.Namespace;
